api/requirements.txt
celery==4.4.0
django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
django_cleanup==4.0.0
djangorestframework==3.11.0
drf-yasg==1.17.0
django-extensions==2.2.5
psycopg2==2.8.3
redis==3.3.11
uwsgi==2.0.18
django-silk==4.0.0
sentry-sdk==0.7.10
coverage==4.5.1
pytest==3.9.3
pytest-django==3.4.4
PyYAML==5.3
tqdm==4.43.0
